refactor : RecommendService의 페이지 조회 switch문 -> enum타입을 이용한 다형성으로 변경
